First step refactor of queue.
* (Partial) split of implementation and interface.
* Remove disfunctioneel confirms from queue.
* Use transient messages for increased performance.
